+++ src/SpinGlassPEPS.jl
@@ -14,15 +14,15 @@ module SpinGlassPEPS
 
     include("base.jl")
     include("utils.jl")
+    include("exact.jl")
     include("compressions.jl")
     include("contractions.jl")
     include("lattice.jl")
-    #include("graphs/chimera.jl")
-    #include("graphs/lattice.jl")
-    include("graph.jl")
     include("ising.jl")
+    include("factor.jl")
+    include("network.jl")
     include("PEPS.jl")
-    include("spectrum.jl")
+    include("MPS_search.jl")
     include("notation.jl")
     include("peps_no_types.jl")
     include("mps_implementation.jl")
