+++ src/factor.jl
@@ -1,84 +1,5 @@
 export factor_graph, projectors
-export Cluster, rank_reveal
-
-const SimpleEdge = LightGraphs.SimpleGraphs.SimpleEdge
-const EdgeIter = Union{LightGraphs.SimpleGraphs.SimpleEdgeIter, Base.Iterators.Filter, Array}
-
-mutable struct Cluster
-    tag::Int
-    vertices::Dict{Int, Int}
-    edges::EdgeIter
-    rank::Vector
-    J::Matrix{<:Number}
-    h::Vector{<:Number}
-
-    function Cluster(ig::MetaGraph, v::Int)
-        cl = new(v)
-        active = filter_vertices(ig, :active, true)
-
-        if cl.tag == 0
-            vlist = vertices(ig)
-        else
-            vlist = filter_vertices(ig, :cell, v)
-        end
-        vlist = intersect(active, vlist)
-
-        L = length(vlist)
-        cl.h = zeros(L)
-        cl.J = zeros(L, L)
-
-        cl.vertices = Dict()
-        cl.edges = SimpleEdge[]
-
-        rank = get_prop(ig, :rank)
-        cl.rank = zeros(Int, L)
-
-        for (i, w) ∈ enumerate(vlist)
-            push!(cl.vertices, w => i)
-            @inbounds cl.h[i] = get_prop(ig, w, :h)
-            @inbounds cl.rank[i] = rank[w]
-        end
-
-        for e ∈ edges(ig)
-            if src(e) ∈ vlist && dst(e) ∈ vlist
-                i, j = cl.vertices[src(e)], cl.vertices[dst(e)]
-                @inbounds cl.J[i, j] = get_prop(ig, e, :J)
-                push!(cl.edges, e)
-            end
-        end
-        cl
-    end
-end
-
-function MetaGraphs.filter_edges(ig::MetaGraph, v::Cluster, w::Cluster)
-    edges = SimpleEdge[]
-    for i ∈ keys(v.vertices), j ∈ neighbors(ig, i)
-        if j ∈ keys(w.vertices) push!(edges, SimpleEdge(i, j)) end
-    end
-    edges
-end
-
-mutable struct Edge
-    tag::NTuple{2, Int}
-    edges::EdgeIter
-    J::Matrix{<:Number}
-
-    function Edge(ig::MetaGraph, v::Cluster, w::Cluster)
-        ed = new((v.tag, w.tag))
-        ed.edges = filter_edges(ig, v, w)
-
-        m = length(v.vertices)
-        n = length(w.vertices)
-
-        ed.J = zeros(m, n)
-        for e ∈ ed.edges
-            i = v.vertices[src(e)]
-            j = w.vertices[dst(e)]
-            @inbounds ed.J[i, j] = get_prop(ig, e, :J)
-        end
-        ed
-    end
-end
+export rank_reveal
 
 function _max_cell_num(ig::MetaGraph)
     L = 0
