--- src/graphs/chimera.jl
@@ -1,66 +1,0 @@
-export Chimera
-export unit_cell
-
-struct Chimera <: Model
-    size::NTuple{3, Int}
-    graph::MetaGraph
-    
-    function Chimera(size::NTuple{3, Int}, graph::MetaGraph)
-        cg = new(size, graph)
-        m, n, t = size
-        linear = LinearIndices((1:m, 1:n))
-
-        for i=1:m, j=1:n, u=1:2, k=1:t
-            v = cg[i, j, u, k]
-            ij = linear[i, j]
-            set_prop!(cg, v, :cell, ij)
-        end
-
-        for e ∈ edges(cg)
-            v = get_prop(cg, src(e), :cell)
-            w = get_prop(cg, dst(e), :cell)
-            set_prop!(cg, e, :cells, (v, w))
-        end
-        cg
-    end
-end
-
-function Chimera(m::Int, n::Int=m, t::Int=4)
-    max_size = m * n * 2 * t
-    g = MetaGraph(max_size)
-
-    hoff = 2t
-    voff = n * hoff
-    mi = m * voff
-    ni = n * hoff
-
-    for i=1:hoff:ni, j=i:voff:mi, k0=j:j+t-1, k1=j+t:j+2t-1
-        add_edge!(g, k0, k1)
-    end
-
-    for i=t:2t-1, j=i:hoff:ni-hoff-1, k=j+1:voff:mi-1
-        add_edge!(g, k, k+hoff-1)
-    end
-
-    for i=1:t, j=i:hoff:ni-1, k=j:voff:mi-voff-1
-        add_edge!(g, k, k+voff)
-    end
-    Chimera((m, n, t), g)
-end
-
-@inline function Base.getindex(c::Chimera, i::Int, j::Int, u::Int, k::Int)
-    _, n, t = size(c)
-    t * (2 * (n * (i - 1) + j - 1) + u - 1) + k
-end
-
-@inline function Base.getindex(c::Chimera, i::Int, j::Int)
-    t = size(c, 3)
-    idx = vec([c[i, j, u, k] for u=1:2, k=1:t])
-    c.graph[idx]
-end
-
-function unit_cell(c::Chimera, v::Int)
-    elist = filter_edges(c.graph, :cells, (v, v))
-    vlist = filter_vertices(c.graph, :cell, v)
-    Cluster(c.graph, v, enum(vlist), elist)
-end
