--- src/graphs/model.jl
@@ -1,31 +1,0 @@
-export Model
-
-abstract type Model end
-
-for op ∈ [
-    :nv,
-    :ne,
-    :eltype,
-    :edgetype,
-    :vertices,
-    :edges,
-    ]
-
-    @eval LightGraphs.$op(c::Model) = $op(c.graph)
-end
-
-for op ∈ [
-    :get_prop,
-    :set_prop!,
-    :has_vertex,
-    :inneighbors,
-    :outneighbors,
-    :neighbors]
-
-    @eval MetaGraphs.$op(m::Model, args...) = $op(m.graph, args...)
-end
-
-#has_edge(m::Model, x...) = has_edge(m.graph, x...)
-
-Base.size(m::Model) = m.size
-Base.size(m::Model, i::Int) = m.size[i]
