+++ src/ising.jl
@@ -1,54 +1,19 @@
 export ising_graph, update_cells!
-export energy, gibbs_tensor
+export energy, Cluster, rank_vec, Spectrum
 
 const Instance = Union{String, Dict}
+const SimpleEdge = LightGraphs.SimpleGraphs.SimpleEdge
+const EdgeIter = Union{LightGraphs.SimpleGraphs.SimpleEdgeIter, Base.Iterators.Filter, Array}
 
-"""
-$(TYPEDSIGNATURES)
-
-Calculates Gibbs state of a classical Ising Hamiltonian
-
-# Details
-
-Calculates matrix elements (probabilities) of \$\\rho\$
-```math
-\$\\bra{\\σ}\\rho\\ket{\\sigma}\$
-```
-for all possible configurations \$\\σ\$.
-"""
-function gibbs_tensor(ig::MetaGraph, β=Float64=1.0)
-    states = collect.(all_states(rank_vec(ig)))
-    ρ = exp.(-β .* energy.(states, Ref(ig)))
-    ρ ./ sum(ρ)
+struct Spectrum
+    energies::Array{<:Number}
+    states::Array{Vector{<:Number}}
 end
 
-"""
-$(TYPEDSIGNATURES)
-
-Calculate the Ising energy
-```math
-E = -\\sum_<i,j> s_i J_{ij} * s_j - \\sum_j h_i s_j.
-```
-"""
-
-energy(σ::Vector, J::Matrix, η::Vector=σ) = dot(σ, J, η)
-energy(σ::Vector, h::Vector) = dot(h, σ)
-energy(σ::Vector, cl::Cluster, η::Vector=σ) = energy(σ, cl.J, η) + energy(cl.h, σ)
-energy(σ::Vector, ig::MetaGraph) = energy(σ, get_prop(ig, :J)) + energy(σ, get_prop(ig, :h))
-
-function energy(fg::MetaDiGraph, edge::Edge)
-    v, w = edge.tag
-    vSp = get_prop(fg, v, :spectrum).states
-    wSp = get_prop(fg, w, :spectrum).states
-
-    m = prod(size(vSp))
-    n = prod(size(wSp))
-
-    en = zeros(m, n)
-    for (j, η) ∈ enumerate(vec(wSp))
-        en[:, j] = energy.(vec(vSp), Ref(edge.J), Ref(η))
-    end
-    en
+function rank_vec(ig::MetaGraph)
+    rank = get_prop(ig, :rank)
+    L = get_prop(ig, :L)
+    Int[get(rank, i, 1) for i=1:L]
 end
 
 """
@@ -128,25 +93,114 @@ function ising_graph(
     ig
 end
 
-"""
-$(TYPEDSIGNATURES)
-
-Calculate unique neighbors of node \$i\$
-
-# Details
-
-This is equivalent of taking the upper
-diagonal of the adjacency matrix
-"""
-function unique_neighbors(ig::MetaGraph, i::Int)
-    nbrs = neighbors(ig::MetaGraph, i::Int)
-    filter(j -> j > i, nbrs)
-end
-
-
 function update_cells!(ig::MetaGraph; rule::Dict)
     for v ∈ vertices(ig)
          w = get_prop(ig, v, :cell)
         set_prop!(ig, v, :cell, rule[w])
     end
+end
+
+mutable struct Cluster
+    tag::Int
+    vertices::Dict{Int, Int}
+    edges::EdgeIter
+    rank::Vector
+    J::Matrix{<:Number}
+    h::Vector{<:Number}
+
+    function Cluster(ig::MetaGraph, v::Int)
+        cl = new(v)
+        active = filter_vertices(ig, :active, true)
+
+        if cl.tag == 0
+            vlist = vertices(ig)
+        else
+            vlist = filter_vertices(ig, :cell, v)
+        end
+        vlist = intersect(active, vlist)
+
+        L = length(vlist)
+        cl.h = zeros(L)
+        cl.J = zeros(L, L)
+
+        cl.vertices = Dict()
+        cl.edges = SimpleEdge[]
+
+        rank = get_prop(ig, :rank)
+        cl.rank = zeros(Int, L)
+
+        for (i, w) ∈ enumerate(vlist)
+            push!(cl.vertices, w => i)
+            @inbounds cl.h[i] = get_prop(ig, w, :h)
+            @inbounds cl.rank[i] = rank[w]
+        end
+
+        for e ∈ edges(ig)
+            if src(e) ∈ vlist && dst(e) ∈ vlist
+                i, j = cl.vertices[src(e)], cl.vertices[dst(e)]
+                @inbounds cl.J[i, j] = get_prop(ig, e, :J)
+                push!(cl.edges, e)
+            end
+        end
+        cl
+    end
+end
+
+function MetaGraphs.filter_edges(ig::MetaGraph, v::Cluster, w::Cluster)
+    edges = SimpleEdge[]
+    for i ∈ keys(v.vertices), j ∈ neighbors(ig, i)
+        if j ∈ keys(w.vertices) push!(edges, SimpleEdge(i, j)) end
+    end
+    edges
+end
+
+mutable struct Edge
+    tag::NTuple{2, Int}
+    edges::EdgeIter
+    J::Matrix{<:Number}
+
+    function Edge(ig::MetaGraph, v::Cluster, w::Cluster)
+        ed = new((v.tag, w.tag))
+        ed.edges = filter_edges(ig, v, w)
+
+        m = length(v.vertices)
+        n = length(w.vertices)
+
+        ed.J = zeros(m, n)
+        for e ∈ ed.edges
+            i = v.vertices[src(e)]
+            j = w.vertices[dst(e)]
+            @inbounds ed.J[i, j] = get_prop(ig, e, :J)
+        end
+        ed
+    end
+end
+
+"""
+$(TYPEDSIGNATURES)
+
+Calculate the Ising energy
+```math
+E = -\\sum_<i,j> s_i J_{ij} * s_j - \\sum_j h_i s_j.
+```
+"""
+
+energy(σ::Vector, J::Matrix, η::Vector=σ) = dot(σ, J, η)
+energy(σ::Vector, h::Vector) = dot(h, σ)
+energy(σ::Vector, cl::Cluster, η::Vector=σ) = energy(σ, cl.J, η) + energy(cl.h, σ)
+energy(σ::Vector, ig::MetaGraph) = energy(σ, get_prop(ig, :J)) + energy(σ, get_prop(ig, :h))
+
+function energy(fg::MetaDiGraph, edge::Edge)
+    v, w = edge.tag
+    vSp = get_prop(fg, v, :spectrum).states
+    wSp = get_prop(fg, w, :spectrum).states
+
+    m = prod(size(vSp))
+    n = prod(size(wSp))
+
+    en = zeros(m, n)
+    for (j, η) ∈ enumerate(vec(wSp))
+        en[:, j] = energy.(vec(vSp), Ref(edge.J), Ref(η))
+    end
+    en
 end
