+++ test/runtests.jl
@@ -35,7 +35,7 @@ push!(my_tests,
     "utils.jl",
     "contractions.jl",
     "compressions.jl",
-    "ising.jl",
+    #"ising.jl",
     "indexing.jl",
     "searchMPS.jl",
     "spectrum.jl",
